+++ guillotina/tests/test_static.py
@@ -9,6 +9,13 @@ def test_get_static_folder(dummy_guillotina):
 
 async def test_render_static_file(container_requester):
     async with await container_requester as requester:
-        response, status = await requester('GET', '/static/testing.py')
+        response, status = await requester('GET', '/static/tests/teststatic.txt')
         assert status == 200
-        assert type(response) == bytes
+        assert response.decode('utf8').strip() == 'foobar'
+
+
+async def test_render_static_default_directory_file(container_requester):
+    async with await container_requester as requester:
+        response, status = await requester('GET', '/static/tests')
+        assert status == 200
+        assert response.decode('utf8').strip() == 'foobar'